testing way to import to modularize commands
